requirements.txt
annotation==0.0.0
apex==0.9.10dev
beautifulsoup4==4.12.2
boto3==1.26.131
botocore==1.29.131
cchardet==2.1.7
dkouqe.egg==info
ebooklib==0.18
faiss==1.5.3
loguru==0.7.0
networkx==3.1
numpy==1.22.3
numpy==1.24.3
pandas==2.0.1
pickle5==0.0.12
pubmed_parser==0.3.1
PyYAML==6.0
PyYAML==6.0
regex==2022.6.2
Requests==2.30.0
scikit_learn==1.2.2
scipy==1.10.1
sentence_transformers==2.2.2
setuptools==67.7.2
six==1.16.0
spacy==3.5.2
streamlit==1.22.0
tensorflow==2.12.0
torch==2.0.0
torch==2.0.1
tqdm==4.65.0
Unidecode==1.3.6
urllib3==2.0.2
Flask-Cors
pymongo
